refactor: SignUpScreen이 Scaffold를 사용하도록 변경
